trying threading backup

+++ src/sprout/oram/operations/AOutput.java
@@ -9,13 +9,13 @@ public class AOutput {
 	BigInteger sC_sig_P_p;
 	BigInteger sE_sig_P_p;
 	BigInteger data;
-	int j_2;
+	BigInteger j_2;
 
 	AOutput() {
 	}
 
 	AOutput(BigInteger l, BigInteger ct, BigInteger et,
-			BigInteger cp, BigInteger ep, BigInteger d, int j2) {
+			BigInteger cp, BigInteger ep, BigInteger d, BigInteger j2) {
 		Lip1 = l;
 		sC_Ti = ct;
 		sE_Ti = et;
